+++ advanced/kernel/src/main/java/org/neo4j/kernel/impl/nioneo/store/RelationshipTypeRecord.java
@@ -41,7 +41,7 @@ public class RelationshipTypeRecord extends AbstractRecord
 
     public void addTypeRecord( DynamicRecord record )
     {
-        typeRecords.put( record.getId(), record );
+        typeRecords.put( (int) record.getId(), record );
     }
 
     public int getTypeBlock()
